requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic_settings
python-multipart
alembic
psycopg2-binary
passlib[bcrypt]
boto3